+++ debug.sh
@@ -2,6 +2,7 @@
 set -e
 
 ## Settings
+version=1
 workdir="${PWD}"
 debug_dir=debug
 debugfile=Debugfile
@@ -17,7 +18,7 @@ else
 fi
 
 watch_enabled=1
-if [[ "$@" == *"--no-watch"* ]]; then
+if [[ "$@" == *"--disable-watch"* ]]; then
     watch_enabled=
 fi
 
@@ -25,36 +26,39 @@ echo ""
 echo "======================="
 echo "## Vtiger Debug Mode ##"
 echo "======================="
+echo " - Debug script version: ${version}"
 echo " - Preparing development environment..."
 
-## Create build dir
-if [[ ! -d "${build_dir}/${debug_dir}" ]]; then
-    mkdir -p "${build_dir}/${debug_dir}"
-fi
+if [[ -n "${watch_enabled}" ]]; then
+    ## Create build dir
+    if [[ ! -d "${build_dir}/${debug_dir}" ]]; then
+        mkdir -p "${build_dir}/${debug_dir}"
+    fi
 
-## Create debugfile
-if [[ ! -f "${build_dir}/${debug_dir}/${debugfile}" ]]; then
-    echo '# Debugfile' > "${build_dir}/${debug_dir}/${debugfile}"
-    echo '*' >> "${build_dir}/${debug_dir}/${debugfile}"
-fi
+    ## Create debugfile
+    if [[ ! -f "${build_dir}/${debug_dir}/${debugfile}" ]]; then
+        echo '# Debugfile' > "${build_dir}/${debug_dir}/${debugfile}"
+        echo '*' >> "${build_dir}/${debug_dir}/${debugfile}"
+    fi
 
-## Copy all source files on build
-cp -RL "${target_dir}"/* "${build_dir}/${debug_dir}"
-find * -type f -not -path "${debug_dir}/*" > "${build_dir}/${debug_dir}/.debugignore"
-while IFS= read line || [[ -n "${line}" ]]; do
-    file=$(echo ${line} | tr -d '\r')
-    [[ -z "${file}" ]] && continue
-    [[ "${file::1}" == "#" ]] && continue
-    rm -f "${build_dir}/${debug_dir}/${file}"
-done < "${build_dir}/${debug_dir}/.debugignore"
-chmod -R 777 "${build_dir}"
-cd "${build_dir}"
-zip -qq -ro debug.zip "${debug_dir}"
-rm -f /app/debug.zip
-mv debug.zip /app/debug.zip
-chmod 777 /app/debug.zip
-rm -fr "${build_dir}"
-cd ${workdir}
+    ## Copy all source files on build
+    cp -RL "${target_dir}"/* "${build_dir}/${debug_dir}"
+    find * -type f -not -path "${debug_dir}/*" > "${build_dir}/${debug_dir}/.debugignore"
+    while IFS= read line || [[ -n "${line}" ]]; do
+        file=$(echo ${line} | tr -d '\r')
+        [[ -z "${file}" ]] && continue
+        [[ "${file::1}" == "#" ]] && continue
+        rm -f "${build_dir}/${debug_dir}/${file}"
+    done < "${build_dir}/${debug_dir}/.debugignore"
+    chmod -R 777 "${build_dir}"
+    cd "${build_dir}"
+    zip -qq -ro debug.zip "${debug_dir}"
+    rm -f /app/debug.zip
+    mv debug.zip /app/debug.zip
+    chmod 777 /app/debug.zip
+    rm -fr "${build_dir}"
+    cd ${workdir}
+fi
 
 echo " - Confirm your environment is ready before start:"
 
@@ -70,7 +74,7 @@ read -p " -> Is it ready? (y/N) " -n 1 -r
 [[ $REPLY =~ ^[Yy]$ ]] || exit 1
 echo ""
 
-if [[ -z "${watch_enabled}" ]]; then
+if [[ -n "${watch_enabled}" ]]; then
     echo ""
     echo "(3) Extract 'debug.zip' file into your project with another terminal (See: https://github.com/javanile/vtiger-dev/wiki/debug.zip)"
     echo "    Windows: Right-click on file 'Show in Explorer' than right-click on file 'Extract here' (or cmd.exe: tar -xvf debug.zip)"
@@ -110,7 +114,7 @@ process_debugfile () {
 }
 
 ## Files watcher
-if [[ -z "${watch_enabled}" ]]; then
+if [[ -n "${watch_enabled}" ]]; then
     echo ""
     echo "Add your additional settings on 'debug/Debugfile'"
     echo "Watching for debug... (Stop with [Ctrl+C])"
@@ -135,5 +139,6 @@ if [[ -z "${watch_enabled}" ]]; then
         #echo ">>> ${filename}"
     done
 else
+    echo ""
     echo "It is time to debug... (Good job)"
 fi
